+++ accounting/src/__tests__/mockserver/internal/handler/pathgetpetfindbystatus.go
@@ -25,7 +25,7 @@ func pathGetPetFindByStatus(dir *logging.HTTPFileDirectory, rt *tracking.Request
 		case "findPetsByStatusTypes[0]":
 			dir.HandlerFunc("findPetsByStatusTypes", testFindPetsByStatusTypesFindPetsByStatusTypes0)(w, req)
 		default:
-			http.Error(w, "Unknown test: "+test, http.StatusBadRequest)
+			http.Error(w, fmt.Sprintf("Unknown test: %s[%d]", test, count), http.StatusBadRequest)
 		}
 	}
 }
@@ -46,7 +46,7 @@ func testFindPetsByStatusTypesFindPetsByStatusTypes0(w http.ResponseWriter, req 
 		http.Error(w, err.Error(), http.StatusBadRequest)
 		return
 	}
-	respBody := []components.Pet{
+	var respBody []components.Pet = []components.Pet{
 		components.Pet{
 			ID:   types.Int64(10),
 			Name: "doggie",